requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0 
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.12.0
beautifulsoup4==4.9.3
pydantic>=2.6.0
selenium>=4.18.1
webdriver-manager>=4.0.1 